requirements.txt
pandas
numpy
matplotlib
seaborn
nltk
googleapiclient.discovery
googleapiclient.errors
scikit-learn
